+++ source/check.cpp
@@ -93,7 +93,7 @@ int check_main()
     for(int i=1;i<=total_sum;++i)
     {
         for(int j=1;j<=50;++j) scout<<"-";
-        scout<<"\r"<<"#"<<i;
+        scout<<"\r#"<<i;
         if(runned_sum-ac_sum!=0)
         {
             for(int j=1;j<=30-std::to_string(i).size();++j) scout<<"-";