+++ source/judge.cpp
@@ -87,9 +87,10 @@ int judge_main()
         // init file
         if(i.value()["in"].is_null()&&i.value()["out"].is_null()) continue;
         for(int j=1;j<=50;++j) scout<<"-";
-        scout<<"\r"<<"#"<<(std::string)i.key()<<"\n";
+        const std::string data_name=i.key();
+        scout<<"\r#"<<data_name<<"\n";
         ++runned_sum;
-        pat run_dir=default_data_dir/"datas"/std::to_string(runned_sum);
+        const pat run_dir=default_data_dir/"datas"/data_name;
         pat in_file=system_nul,out_file=system_nul,ans_file=run_dir/"data.ans",chk_file=run_dir/"data.txt";
         std::filesystem::create_directory(run_dir);
         if(!i.value()["in"].is_null()) std::filesystem::copy_file((std::string)i.value()["in"],in_file=run_dir/"data.in",std::filesystem::copy_options::overwrite_existing);
@@ -113,7 +114,7 @@ int judge_main()
         if(run_judger.result.istrue()) ++ac_sum;
         if(run_judger.result.isfalse())
         {
-            std::filesystem::copy(run_dir,default_data_dir/(std::to_string(runned_sum)+" - "+get_short_resultname(run_judger.result)),std::filesystem::copy_options::recursive);
+            std::filesystem::copy(run_dir,default_data_dir/(data_name+" - "+get_short_resultname(run_judger.result)),std::filesystem::copy_options::recursive);
         }
         #undef run_dir
     }